requirements.txt
python-dotenv==1.0.1
openai==1.30.1
langchain==0.2.0
langchain-community==0.2.0
langchain-openai==0.1.7